Iter5 (#5)
* implement server

* reduce mod ver

* transform constants to camelcase

* fix erorr cases

* implement collector & client

* add test for update handler

* rm unused types

* clean

* clean&undo

* retur missed server.go

* add chi; add root handler

* fix float

* WIP

* remove acc from storage; use sum in collector

* WIP

* undo precision; set it to 3

* use sum in storage

* WIP

* WIP

* fix missed return

* rm wg

* format float with min possible precision

* rm unused dep

* add flags for server & agent

* edit agent regex

* rm unused mux

* edit addr validation

* add optional use of env vars

* fix setting intervals

* mv html to template file; add template parse, exec in handler

* set prev version of goimports

* refactor server

* add tests for each handler; add negative cases

+++ internal/collector/collector.go
@@ -15,7 +15,7 @@ type Collector struct {
 	pollInterval uint
 }
 
-func New(pollInterval uint8) *Collector {
+func New(pollInterval uint) *Collector {
 	m := Mtcs{
 		Gauge:   make(map[string]float64),
 		Counter: make(map[string]int64),
